Create a throttler for each comm instance; ensure throttled function has no arguments; stylistic tweaks
